cleanup, added crate doc
