+++ src/rabject/rabject2d/mod.rs
@@ -10,7 +10,7 @@ use std::ops::{Deref, DerefMut};
 use log::trace;
 
 use crate::scene::canvas::camera::CanvasCamera;
-use crate::{context::RanimContext, scene::entity::Entity, updater::Updater, utils::Id};
+use crate::{context::RanimContext, scene::Entity, updater::Updater, utils::Id};
 
 use crate::rabject::{Primitive, Rabject};
