+++ src/scene/entity.rs
@@ -1,6 +1,6 @@
-use std::any::Any;
+use std::{any::Any, fmt::Debug, marker::PhantomData, ops::Deref};
 
-use crate::context::RanimContext;
+use crate::{context::RanimContext, utils::Id};
 
 pub trait EntityAny: Entity + Any {
     fn as_any(&self) -> &dyn Any;
@@ -17,9 +17,6 @@ impl<T: Entity + Any> EntityAny for T {
 }
 
 /// An entity in the scene
-///
-/// This only implent way to [`Entity::tick`], [`Entity::extract`], [`Entity::prepare`] the entity.
-/// For the actual render code, check the implementor of [`crate::camera::Renderer<E: EntityAny>`]
 pub trait Entity {
     type Renderer;
 
@@ -28,3 +25,33 @@ pub trait Entity {
     fn prepare(&mut self, ctx: &RanimContext);
     fn render(&mut self, ctx: &mut RanimContext, renderer: &mut Self::Renderer);
 }
+
+pub struct EntityId<E: EntityAny>(Id, PhantomData<E>);
+
+impl<E: EntityAny> Debug for EntityId<E> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "EntityId({:?})", self.0)
+    }
+}
+
+impl<E: EntityAny> Copy for EntityId<E> {}
+
+impl<E: EntityAny> Clone for EntityId<E> {
+    fn clone(&self) -> Self {
+        *self
+    }
+}
+
+impl<E: EntityAny> EntityId<E> {
+    pub fn from_id(id: Id) -> Self {
+        Self(id, PhantomData)
+    }
+}
+
+impl<E: EntityAny> Deref for EntityId<E> {
+    type Target = Id;
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
