+++ src/utils/rate_functions.rs
@@ -4,7 +4,7 @@ pub fn linear(t: f32) -> f32 {
 
 /// Smooth rate function
 ///
-/// from https://github.com/3b1b/manim/blob/003c4d86262565bb21001f74f67e6788cae62df4/manimlib/utils/rate_functions.py#L17
+/// from <https://github.com/3b1b/manim/blob/003c4d86262565bb21001f74f67e6788cae62df4/manimlib/utils/rate_functions.py#L17>
 pub fn smooth(t: f32) -> f32 {
     let s = 1.0 - t;
     t * t * t * (10.0 * s * s + 5.0 * s * t + t * t)
